refactor: move internal functions to a saparate file
